requirements.txt
streamlit>=1.30.0
pymongo>=4.6.0
python-dotenv>=1.0.0
extra-streamlit-components>=0.1.60
